Refactor of GeoShape integration tests (#77052)
This commit joins GeoFilterIT and GeoShapeIntegrationIT into one test case called GeoShapeIntegTestCase
which is moved into the test framework.
# Conflicts:
#	server/src/internalClusterTest/java/org/elasticsearch/search/geo/GeoFilterIT.java
#	server/src/internalClusterTest/java/org/elasticsearch/search/geo/GeoShapeIntegrationIT.java
#	server/src/internalClusterTest/java/org/elasticsearch/search/geo/LegacyGeoShapeIntegrationIT.java
