+++ server/src/internalClusterTest/java/org/elasticsearch/search/geo/GeoShapeIT.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+package org.elasticsearch.search.geo;
+
+import org.elasticsearch.common.xcontent.XContentBuilder;
+
+import java.io.IOException;
+
+public class GeoShapeIT extends GeoShapeIntegTestCase {
+
+    @Override
+    protected void getGeoShapeMapping(XContentBuilder b) throws IOException {
+        b.field("type", "geo_shape");
+    }
+
+    @Override
+    protected boolean allowExpensiveQueries() {
+        return true;
+    }
+}
